+++ src/pseudonymize/detectors/registry.py
@@ -4,6 +4,7 @@ from pseudonymize.detectors.email import EmailDetector
 from pseudonymize.detectors.iban import IbanDetector
 from pseudonymize.detectors.ip_address import IpAddressDetector
 from pseudonymize.detectors.italian import ItalianFiscalCodeDetector, ItalianVATDetector
+from pseudonymize.detectors.location import LocationDetector
 from pseudonymize.detectors.payment_card import PaymentCardDetector
 from pseudonymize.detectors.phone import PhoneDetector
 from pseudonymize.detectors.secret import SecretDetector
@@ -20,4 +21,5 @@ DEFAULT_DETECTORS: tuple[Detector, ...] = (
     PhoneDetector(),
     UrlDetector(),
     SecretDetector(),
+    LocationDetector(),
 )